+++ var/phpstan/cache/PHPStan/12/26/1226faa804108ffb4723f827d27916a162bb8066.php
@@ -1,0 +1,7 @@
+<?php declare(strict_types = 1);
+
+// variadic-method-Sif\Foundation\Contracts\ApplicationInterface-hasCapability-D:\SIF\src\Foundation\Contracts\ApplicationInterface.php
+return \PHPStan\Cache\CacheItem::__set_state(array(
+   'variableKey' => '1784128034-v4',
+   'data' => false,
+));