+++ var/phpstan/cache/PHPStan/29/83/2983831b75d92c8bf4648781fd4157a6d1d7ebca.php
@@ -1,0 +1,7 @@
+<?php declare(strict_types = 1);
+
+// variadic-method-Sif\Foundation\Events\KernelBooting-jsonSerialize-D:\SIF\src\Foundation\Events\KernelBooting.php
+return \PHPStan\Cache\CacheItem::__set_state(array(
+   'variableKey' => '1784128095-v4',
+   'data' => false,
+));