+++ var/phpstan/cache/PHPStan/2f/12/2f129c084ca92af6c7a6ca6544dfbb6f6ce3710b.php
@@ -1,0 +1,7 @@
+<?php declare(strict_types = 1);
+
+// variadic-method-Sif\Foundation\Events\FrameworkBooting-timestamp-D:\SIF\src\Foundation\Events\FrameworkBooting.php
+return \PHPStan\Cache\CacheItem::__set_state(array(
+   'variableKey' => '1784128090-v4',
+   'data' => false,
+));