+++ var/phpstan/cache/PHPStan/46/09/4609b3b2927861b43eb5f02d875b6397d6deb38a.php
@@ -1,0 +1,7 @@
+<?php declare(strict_types = 1);
+
+// variadic-method-Sif\Foundation\Tests\Unit\RuntimeEventTest-applicationEvents-D:\SIF\tests\Foundation\Unit\RuntimeEventTest.php
+return \PHPStan\Cache\CacheItem::__set_state(array(
+   'variableKey' => '1784128196-v4',
+   'data' => false,
+));