+++ var/phpstan/cache/PHPStan/8b/a5/8ba5b77fcd4da02a890c1afb9b669c2cdc3c922d.php
@@ -1,0 +1,7 @@
+<?php declare(strict_types = 1);
+
+// variadic-method-Sif\Foundation\Events\ApplicationBooted-timestamp-D:\SIF\src\Foundation\Events\ApplicationBooted.php
+return \PHPStan\Cache\CacheItem::__set_state(array(
+   'variableKey' => '1784128094-v4',
+   'data' => false,
+));