+++ var/phpstan/cache/PHPStan/b8/28/b8281d9b72550063ef8f3c0a5c5be93252003ee3.php
@@ -2,6 +2,6 @@
 
 // variadic-method-Sif\Foundation\Application-__construct-D:\SIF\src\Foundation\Application.php
 return \PHPStan\Cache\CacheItem::__set_state(array(
-   'variableKey' => '1784124133-v4',
+   'variableKey' => '1784128035-v4',
    'data' => false,
 ));