+++ var/phpstan/cache/PHPStan/d1/2c/d12c55f9f899bfc4a7b3e5297edfe9db4e8c2481.php
@@ -1,0 +1,7 @@
+<?php declare(strict_types = 1);
+
+// variadic-method-Sif\Foundation\Events\FrameworkBooted-application-D:\SIF\src\Foundation\Events\FrameworkBooted.php
+return \PHPStan\Cache\CacheItem::__set_state(array(
+   'variableKey' => '1784128091-v4',
+   'data' => false,
+));