+++ var/phpstan/cache/PHPStan/d2/56/d256281645918c6c8574cce1c6d3830090393859.php
@@ -1,0 +1,7 @@
+<?php declare(strict_types = 1);
+
+// variadic-method-Sif\Foundation\Events\KernelBooted-jsonSerialize-D:\SIF\src\Foundation\Events\KernelBooted.php
+return \PHPStan\Cache\CacheItem::__set_state(array(
+   'variableKey' => '1784128097-v4',
+   'data' => false,
+));