+++ var/phpstan/cache/PHPStan/eb/03/eb038e21717ae9f08bcea2d303b091120d5cdec2.php
@@ -1,0 +1,7 @@
+<?php declare(strict_types = 1);
+
+// variadic-method-Sif\Foundation\Exceptions\InvalidCapabilityException-empty-D:\SIF\src\Foundation\Exceptions\InvalidCapabilityException.php
+return \PHPStan\Cache\CacheItem::__set_state(array(
+   'variableKey' => '1784128036-v4',
+   'data' => false,
+));